tips on leftbar and topbar
