+++ client/src/components/UI/navbar/LeftBar/LeftBar.js
@@ -1,9 +1,8 @@
 import "./style.css";
 import { useState, useEffect } from "react";
-import Tooltip, { tooltipClasses } from '@mui/material/Tooltip';
 import OrangeLogo from "../../svg/OrangeLogo";
-import { styled } from '@mui/material/styles';
 import IconButton from '@mui/material/IconButton';
+import LightToolTip from '../../tip/LightToolTip'
 
 // Icones do MATERIAL ICONS
 
@@ -15,16 +14,6 @@ import NotificationsIcon from '@mui/icons-material/Notifications';
 import PostAddIcon from '@mui/icons-material/PostAdd';
 import TrendingUpIcon from '@mui/icons-material/TrendingUp';
 import CalendarMonthIcon from '@mui/icons-material/CalendarMonth';
-
-const LightTooltip = styled(({ className, ...props }) => (
-  <Tooltip {...props} classes={{ popper: className }} />
-))(({ theme }) => ({
-  [`& .${tooltipClasses.tooltip}`]: {
-    backgroundColor: '#654C8F',
-    color: 'white',
-    fontSize: 11,
-  },
-}));
 
 export default function LeftBar () {
 
@@ -72,9 +61,9 @@ export default function LeftBar () {
 
             {!isLoading ? (
               <>
-              <LightTooltip placement="right" followCursor title="Painel de controle">
+              <LightToolTip placement="right" followCursor title="Painel de controle">
                 <HomeIcon alt='homepage' sx={{ width: 28, height: 28, color: "white" }} />
-              </LightTooltip>
+              </LightToolTip>
               
               <p>Painel de controle</p>
               
@@ -89,9 +78,9 @@ export default function LeftBar () {
 
           {!isLoading ? (
               <>
-              <LightTooltip placement="right" followCursor title="Desempenho escolar">
+              <LightToolTip placement="right" followCursor title="Desempenho escolar">
                 <TrendingUpIcon alt='desempenho' sx={{ width: 28, height: 28, color: "white" }} />
-              </LightTooltip>
+              </LightToolTip>
               
               <p>Desempenho</p>
 
@@ -110,9 +99,9 @@ export default function LeftBar () {
 
           {!isLoading ? (
               <>
-              <LightTooltip placement="right" followCursor title="Calendário de notas">
+              <LightToolTip placement="right" followCursor title="Calendário de notas">
                 <CalendarMonthIcon alt='calendario de notas' sx={{ width: 28, height: 28, color: "white" }} />
-              </LightTooltip>
+              </LightToolTip>
               <p>Calendario de Notas</p>
 
               </>
@@ -124,9 +113,9 @@ export default function LeftBar () {
           <div className="iconDiv">
           {!isLoading ? (
               <>
-              <LightTooltip placement="right" followCursor title="Painel de notificações">
+              <LightToolTip placement="right" followCursor title="Painel de notificações">
                 <NotificationsIcon alt='notificações' sx={{ width: 28, height: 28, color: "white"  }} />
-              </LightTooltip>
+              </LightToolTip>
 
               <p>Avisos e notificações</p>
 
@@ -141,9 +130,9 @@ export default function LeftBar () {
 
           {!isLoading ? (
               <>
-              <LightTooltip placement="right" followCursor title="Regulamento escolar">
+              <LightToolTip placement="right" followCursor title="Regulamento escolar">
                 <PostAddIcon alt='regulamento escolar' sx={{ width: 28, height: 28, color: "white" }} />
-              </LightTooltip>
+              </LightToolTip>
 
               <p>Regulamento escolar</p>
 
@@ -162,9 +151,9 @@ export default function LeftBar () {
           <div className="iconDiv">
           {!isLoading ? (
               <>
-              <LightTooltip placement="right" followCursor title="Tarefas">
+              <LightToolTip placement="right" followCursor title="Tarefas">
                 <School alt='Tarefas' sx={{ width: 28, height: 28, color:"white" }} />
-              </LightTooltip>
+              </LightToolTip>
                 <p>Lista de atividades</p>
 
               </>
@@ -178,9 +167,9 @@ export default function LeftBar () {
               <Skeleton variant="circular"  sx={{ bgcolor: '#310A6E', marginTop:'2.3rem'  }} width={28} height={28} />
             ) : (
               <>
-              <LightTooltip placement="right" followCursor title="Averiguação do sistema">
+              <LightToolTip placement="right" followCursor title="Averiguação do sistema">
                 <Settings sx={{ width: 28, height: 28, color:"white" }} />
-              </LightTooltip>
+              </LightToolTip>
                 <p>Averiguação do sistema</p>
               </>
             )}
