+++ client/src/components/UI/tip/LightToolTip.js
@@ -1,0 +1,14 @@
+import { styled } from '@mui/material/styles';
+import Tooltip, { tooltipClasses } from '@mui/material/Tooltip';
+
+const LightTooltip = styled(({ className, ...props }) => (
+    <Tooltip {...props} classes={{ popper: className }} />
+  ))(({ theme }) => ({
+    [`& .${tooltipClasses.tooltip}`]: {
+      backgroundColor: '#654C8F',
+      color: 'white',
+      fontSize: 11,
+    },
+  }));
+
+  export default LightTooltip;
